# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MDDB API
  description: |
    MDDB (Markdown Database) is a high-performance, version-controlled markdown database with triple protocol support.

    ## Features
    - Store and manage markdown documents with metadata
    - Full revision history for all documents
    - Multi-language support
    - Fast metadata-based search
    - **Vector search** with embedding providers (OpenAI, Cohere, Voyage AI, Ollama)
    - **Full-text search** with TF-IDF, BM25, and BM25F field-weighted scoring
    - **Hybrid search** combining FTS + vector search with alpha blending or RRF fusion
    - **Webhooks** for real-time event notifications
    - **TTL** (time-to-live) support for automatic document expiration
    - **JSON Schema validation** for document metadata
    - **URL import** with YAML frontmatter parsing
    - **Authentication** - JWT tokens and API keys
    - **Authorization** - Collection-level RBAC (Read/Write/Admin)
    - **User Management** - Multi-user with admin roles
    - **Group Permissions** - Organize users into groups
    - Template variable substitution
    - Export capabilities (NDJSON, ZIP)
    - Backup and restore operations
    - **Replication** - Leader-follower replication with binlog streaming

    ## Protocols
    - **HTTP/JSON** - Port 11023 (documented here)
    - **gRPC/Protobuf** - Port 11024 (see gRPC documentation)
    - **GraphQL** - Port 11023/graphql (enable with MDDB_GRAPHQL_ENABLED=true)
...

    ## Authentication
...

...

...
    ```

...
servers:
...
paths:
...
  /v1/automation:
    get:
      tags: [Automation]
      summary: List automation rules
      description: Returns all automation rules (webhooks, triggers, crons). Optionally filter by type.
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [webhook, trigger, cron]
          description: Filter by rule type
      responses:
        "200":
          description: List of automation rules
          content:
            application/json:
              schema:
                type: object
                properties:
                  rules:
                    type: array
                    items:
                      $ref: '#/components/schemas/AutomationRule'
                  total:
                    type: integer
    post:
      tags: [Automation]
      summary: Create automation rule
      description: Create a new webhook target, trigger, or cron rule.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutomationRule'
      responses:
        "201":
          description: Rule created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationRule'
        "400":
          description: Validation error
  /v1/automation/{id}:
    get:
      tags: [Automation]
      summary: Get automation rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Automation rule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationRule'
        "404":
          description: Not found
    put:
      tags: [Automation]
      summary: Update automation rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutomationRule'
      responses:
        "200":
          description: Rule updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationRule'
    delete:
      tags: [Automation]
      summary: Delete automation rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Rule deleted
...
components:
  schemas:
...
    AutomationRule:
      type: object
      required: [type, name]
      properties:
        id:
          type: string
          readOnly: true
        type:
          type: string
          enum: [webhook, trigger, cron]
        name:
          type: string
        enabled:
          type: boolean
          default: true
        createdAt:
          type: integer
          readOnly: true
        updatedAt:
          type: integer
          readOnly: true
        url:
          type: string
          description: Webhook URL (type=webhook)
        method:
          type: string
          enum: [POST, GET, PUT]
          default: POST
          description: HTTP method (type=webhook)
        headers:
          type: object
          additionalProperties:
            type: string
          description: Custom HTTP headers (type=webhook)
        collection:
          type: string
          description: Target collection (type=trigger)
        searchType:
          type: string
          enum: [fts, vector, hybrid]
          description: Search type (type=trigger)
        query:
          type: string
          description: Search query (type=trigger)
        threshold:
          type: number
          minimum: 0
          maximum: 100
          description: Minimum score threshold 0-100 (type=trigger)
        webhookId:
          type: string
          description: Target webhook ID (type=trigger)
        searchParams:
          type: object
          description: Extra search parameters (type=trigger)
        schedule:
          type: string
          description: Cron expression e.g. "0 9 * * *" (type=cron)
        triggerId:
          type: string
          description: Target trigger ID (type=cron)
        lastRun:
          type: integer
          readOnly: true
          description: Last execution timestamp (type=cron)
        nextRun:
          type: integer
          readOnly: true
          description: Next scheduled execution (type=cron)
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /v1/auth/login
    APIKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key obtained from POST /v1/auth/api-key